Change logic around saving and publishing
There are three states to consider: the state of the form, the state
of the preview, and the saved state (at least as we last remember it).
Before, we allowed publishing when the preview state matched the saved
state.  But this produces odd configurations.  If the form state has
changed, but the preview is still the same as the saved state, hitting
"Save" causes removes the ability to publish (since it chanes the saved
state).  Or, if all three states are different, the user has to hit both
"Save" and "Preview" to be able to publish.

By autosaving on Publish, we can perform one of those syncs.  But the
PDF that we have corresponds to the preview state, which doesn't enter
into the save system.  In order to make sure the PDF corresponds to the
saved state, which is being set to the form state, we now only allow
publishing when the form state equals the preview state.  The publish
action (now "Save and Publish") will synchronize both the saved JSON
and PDF to this state.

I hope.